moved geom tests to a separate util dir
