+++ .pre-commit-config.yaml
@@ -11,7 +11,6 @@ repos:
     - id: flake8
       args:
         - "--max-line-length=120"
-        - "--ignore=F405, E712, F403"
 - repo: https://github.com/pycqa/isort
   rev: 5.11.5
   hooks:
