+++ pyproject.toml
@@ -49,15 +49,15 @@ skip = [
 ]
 
 [tool.mypy]
-strict_optional = true
-warn_return_any = true
-warn_redundant_casts = true
-check_untyped_defs = true
-disallow_any_generics = true
-no_implicit_optional = true
-warn_unused_configs = true
-show_error_codes = true
-show_column_numbers = true
+strict_optional = true # no implicit Optional[Any] for None
+warn_return_any = true # if return type is any for a function
+warn_redundant_casts = true # if a cast to a type is unnecessary
+check_untyped_defs = true # warn if a function was not annotated with types
+disallow_any_generics = true # warn if generics are used
+no_implicit_optional = true # warn on non-optional usage of variables initialized as None
+warn_unused_configs = true # warn on unused imports
+show_error_codes = true # show error codes in error messages
+show_column_numbers = true # show column numbers in error messages
 exclude = [
     '\.git',
     '__pycache__',
