+++ requirements.txt
@@ -3,3 +3,4 @@ requests==2.20.0 ; python_version >= "3.7" and python_full_version < "4.0.0"
 typer[all]==0.9.0 ; python_version >= "3.7" and python_full_version < "4.0.0"
 halo==0.0.29 ; python_version >= "3.7" and python_full_version < "4.0.0"
 pre-commit==2.20.0 ; python_version >= "3.7" and python_full_version < "4.0.0"
+pytest==7.0.0 ; python_version >= "3.7" and python_full_version < "4.0.0"
